requirements.txt
flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.2
openai==1.3.0
redis==5.0.1
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
markdown==3.5.1
gunicorn==21.2.0 